+++ app/src/test/java/kotlin/coroutines/Continuation.java
@@ -1,4 +1,0 @@
-package kotlin.coroutines;
-
-public interface Continuation {
-}